Issue201 - just move a a func that's only used in one spot
See issue #201 for history - just seems unnecessary.
